refactor: move DOMPurify functions to server-only and keep shared/utils/html lightweight
